[halium-wrappers] Add libgtk6216workaround.so to workaround gtk glyphs bug
GTK 4.12.4+ expects that the underlying GLES implementation would
convert textures if a different format is passed to glTexSubImage2D.

This actually works in Mesa, but it's out of spec, and breaks
glyphs rendering in other implementations (such as NVIDIA, and at
least some Adreno devices).

This workaround tells GTK that the BGRA8888 format is not supported
(note: it's actually supported in most of the devices), so that it
can fallback to RGBA which matches the original texture format.

Only users of libepoxy are affected by this workaround.

Bug reference: https://gitlab.gnome.org/GNOME/gtk/-/issues/6216

This workaround is shipped as part of the new `libdroidian-workarounds` and
`flatpak-extension-gl-hybris-droidian-workarounds` packages.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>