[halium-wrappers] Add libglesshadercache.so to improve app startup time
On many devices, we spend between 1 and 3 seconds compiling and linking shaders
every time an app is started. This especially affects GTK4 apps, which compiles
and links 51 different programs (!) on each startup.

The cache can be shared between different apps, since we use a checksum of all
the shaders involved in each shader program as the filename. This means that
even newly-downloaded software that also uses GTK4 will be faster to start up.

Unfortunately, the only way to share the shader cache between Flatpak apps is by
applying a nasty hack to bind a directory as R/W in every Flatpak. So we don't
share caches between Flatpaks. At any rate, this might be good for security:
most graphics drivers don't offer any security guarantees when loading shader
binaries.

Signed-off-by: Jesús Higueras <jesus@dabbleam.com>